moving ConfirmationText into ShowIdentity
