--- src/components/display/fewlines/ConfirmationBox/ConfirmationText.tsx
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const ConfirmationText = styled.p`
-  margin: 0 0 ${({ theme }) => theme.spaces.component.xs};
-`;
